Move ScanFilterInterface to internal
