+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/scan/ScanResult.java
@@ -3,6 +3,7 @@ package com.polidea.rxandroidble2.scan;
 import androidx.annotation.NonNull;
 
 import com.polidea.rxandroidble2.RxBleDevice;
+import com.polidea.rxandroidble2.internal.ScanResultInterface;
 import com.polidea.rxandroidble2.internal.logger.LoggerUtil;
 
 public class ScanResult implements ScanResultInterface {
